backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gtscr"
version = "1.0.4"
description = "A tool to generate commit messages using OpenAI models."
readme = "README.md"
authors = [
    { name = "GitScribe Team" }
]
requires-python = ">=3.12"

dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
    "fastapi>=0.117.1",
    "python-multipart>=0.0.9",
    "GitPython>=3.1.46"
]

[project.scripts]
gitscribe = "gitscribe.main:main"
gs = "gitscribe.main:main"

[tool.setuptools]
packages = ["gitscribe"]